expand os for c++ testing
