+++ src/cpp/base.cpp
@@ -14,8 +14,7 @@
 
 namespace perspective {
 
-void
-psp_abort() {
+void psp_abort() {
     std::raise(SIGINT);
 }
 
@@ -315,6 +314,14 @@ bool
 is_internal_colname(const std::string& c) {
     return c.compare(std::string("psp_")) == 0;
 }
+
+
+template <typename T>
+t_dtype
+type_to_dtype() {
+    return DTYPE_NONE;
+}
+
 
 template <>
 t_dtype
