+++ src/cpp/context_grouped_pkey.cpp
@@ -29,6 +29,12 @@ t_ctx_grouped_pkey::t_ctx_grouped_pkey()
     : m_depth(0),
     m_depth_set(false) {}
 
+t_ctx_grouped_pkey::t_ctx_grouped_pkey(t_schema schema, t_config config)
+	: m_depth(0),
+	m_depth_set(false) {
+	PSP_COMPLAIN_AND_ABORT('Not Implemented');
+}
+
 t_ctx_grouped_pkey::~t_ctx_grouped_pkey() {}
 
 void
