+++ src/cpp/sparse_tree.cpp
@@ -27,7 +27,7 @@
 #include <perspective/table.h>
 #include <perspective/filter_utils.h>
 #include <perspective/context_two.h>
-#include <unordered_set>
+#include <set>
 
 namespace perspective {
 
@@ -1011,13 +1011,13 @@ t_stree::update_agg_table(t_uindex nidx, t_agg_update_info& info, t_uindex src_r
                 new_value.set(gstate.reduce<std::function<t_tscalar(std::vector<t_tscalar>&)>>(
                     pkeys, spec.get_dependencies()[0].name(),
                     [this](std::vector<t_tscalar>& values) {
-                        std::unordered_set<t_tscalar> vset;
+                        std::set<t_tscalar> vset;
                         for (const auto& v : values) {
                             vset.insert(v);
                         }
 
                         std::stringstream ss;
-                        for (std::unordered_set<t_tscalar>::const_iterator iter = vset.begin();
+                        for (std::set<t_tscalar>::const_iterator iter = vset.begin();
                              iter != vset.end(); ++iter) {
                             ss << *iter << ", ";
                         }
