+++ src/include/perspective/context_grouped_pkey.h
@@ -21,7 +21,8 @@ namespace perspective {
 
 class PERSPECTIVE_EXPORT t_ctx_grouped_pkey : public t_ctxbase<t_ctx_grouped_pkey> {
 public:
-    t_ctx_grouped_pkey();
+	t_ctx_grouped_pkey();
+	t_ctx_grouped_pkey(t_schema schema, t_config config);
 
     ~t_ctx_grouped_pkey();
