+++ src/include/perspective/first.h
@@ -8,7 +8,9 @@
  */
 
 #ifndef PSP_ENABLE_WASM
+#ifndef PSP_PARALLEL_FOR
 #define PSP_PARALLEL_FOR
+#endif
 #endif
 
 #if !defined(__linux__) && !defined(__APPLE__) && !defined(WIN32)
