+++ src/include/perspective/table.h
@@ -160,7 +160,7 @@ private:
     bool m_from_recipe;
 };
 
-bool operator==(const t_table& lhs, const t_table& rhs);
+PERSPECTIVE_EXPORT bool operator==(const t_table& lhs, const t_table& rhs);
 
 template <typename FLATTENED_T>
 void
