+++ test/CMakeLists.txt
@@ -6,5 +6,9 @@ set (TEST_FILES
 include_directories("include")
 add_executable(psp_test ${TEST_FILES})
 
-target_link_libraries(psp_test gtest_main tbb psp)
+if(WIN32)
+	target_compile_definitions(psp_test PRIVATE WIN32=1)
+endif()
+
+target_link_libraries(psp_test psp gtest_main tbb )
 add_test(NAME psptest COMMAND psp_test)
